requirements.txt
pandas>=2.0.0
openpyxl>=3.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
xlrd>=2.0.0
